requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
eventlet==0.36.1
rapidfuzz==3.9.7
vosk==0.3.45
pyttsx3==2.90
reportlab==4.2.2
noisereduce==3.0.2

opencv-python==4.10.0.84
# Let pip resolve a compatible wheel for your Python/Windows environment
mediapipe
numpy==1.26.4
requests==2.32.3
